+++ Projects/DonWorry/Sources/Home/Cells/GiveBillCardCollectionViewCell.swift
@@ -1,5 +1,5 @@
 //
-//  TakePaymentCardCollectionViewCell.swift
+//  GiveBillCardCollectionViewCell.swift
 //  DonWorry
 //
 //  Created by Woody on 2022/08/09.
@@ -8,23 +8,49 @@
 
 import UIKit
 import DesignSystem
+import DonWorryExtensions
+import Kingfisher
 
-struct TakePaymentCardCellViewModel: Equatable {
-    var giverID: Int
+struct GiveBillCardCellViewModel: Equatable {
+    var takerID: Int
+    var imageURL: String
+    var nickName: String
     var amount: String
     var isCompleted: Bool
 }
 
-final class TakePaymentCardCollectionViewCell: UICollectionViewCell {
+final class GiveBillCardCollectionViewCell: UICollectionViewCell {
 
-    lazy var titleLabel: UILabel = {
-        let v = UILabel()
-        v.font = .designSystem(weight: .bold, size: ._13)
-        v.textColor = .designSystem(.white)
-        v.text = "받을 돈"
+    lazy var wholeStackView: UIStackView = {
+        let v = UIStackView()
+        v.axis = .vertical
+        v.alignment = .center
+        v.distribution = .equalSpacing
         return v
     }()
 
+    lazy var profileStackView: UIStackView = {
+        let v = UIStackView()
+        v.axis = .vertical
+        v.alignment = .center
+        v.distribution = .fill
+        v.spacing = 10
+        return v
+    }()
+    lazy var profileImageView: UIImageView = {
+        let v = UIImageView()
+        v.layer.borderColor = UIColor.designSystem(.white)?.cgColor
+        v.layer.borderWidth = 1
+        v.contentMode = .scaleAspectFill
+        return v
+    }()
+    lazy var nickNameLabel: UILabel = {
+        let v = UILabel()
+        v.textColor = .designSystem(.white)
+        v.font = .designSystem(weight: .heavy, size: ._15)
+        v.textAlignment = .center
+        return v
+    }()
     lazy var descriptionStackView: UIStackView = {
         let v = UIStackView()
         v.axis = .vertical
@@ -37,7 +63,7 @@ final class TakePaymentCardCollectionViewCell: UICollectionViewCell {
         let v = UILabel()
         v.textColor = .designSystem(.white)
         v.font = .designSystem(weight: .bold, size: ._13)
-        v.text = "미정산"
+        v.text = "줄돈"
         v.textAlignment = .center
         return v
     }()
@@ -54,9 +80,14 @@ final class TakePaymentCardCollectionViewCell: UICollectionViewCell {
         return v
     }()
 
-    var viewModel: TakePaymentCardCellViewModel? {
+    var viewModel: GiveBillCardCellViewModel? {
         didSet {
+            self.nickNameLabel.text = viewModel?.nickName
             self.amountLabel.text = viewModel?.amount
+            if let viewModel = viewModel {
+                let urlString = URL(string: viewModel.imageURL)
+                profileImageView.kf.setImage(with: urlString)
+            }
             if let viewModel = viewModel {
                 completeCoverView.isHidden = !viewModel.isCompleted
             }
@@ -75,23 +106,29 @@ final class TakePaymentCardCollectionViewCell: UICollectionViewCell {
     }
 
     private func setUI() {
-        self.contentView.backgroundColor = .designSystem(.green)
+        self.contentView.backgroundColor = .designSystem(.brown)
+        self.contentView.addSubview(self.wholeStackView)
         self.contentView.addSubview(self.completeCoverView)
-        self.contentView.addSubview(self.titleLabel)
-        self.contentView.addSubview(self.descriptionStackView)
+        self.wholeStackView.addArrangedSubview(self.profileStackView)
+        self.wholeStackView.addArrangedSubview(self.descriptionStackView)
+        self.profileStackView.addArrangedSubview(self.profileImageView)
+        self.profileStackView.addArrangedSubview(self.nickNameLabel)
         self.descriptionStackView.addArrangedSubview(self.descriptionLabel)
         self.descriptionStackView.addArrangedSubview(self.amountLabel)
 
-        self.titleLabel.snp.makeConstraints { make in
-            make.top.equalToSuperview().offset(15)
-            make.leading.equalToSuperview().offset(14)
+        self.wholeStackView.snp.makeConstraints { make in
+            make.top.equalToSuperview().offset(25)
+            make.bottom.equalToSuperview().inset(19)
+            make.leading.trailing.equalToSuperview().inset(25)
         }
-        self.descriptionStackView.snp.makeConstraints { make in
-            make.center.equalToSuperview()
+        self.profileImageView.snp.makeConstraints { make in
+            make.width.height.equalTo(60)
         }
         self.completeCoverView.snp.makeConstraints { make in
             make.edges.equalToSuperview()
         }
+
+        self.profileImageView.roundCorners(30)
         self.contentView.roundCorners(8)
         self.addShadow(shadowColor: UIColor.black.withAlphaComponent(0.4).cgColor)
         self.addCompleteCoverViewBlurEffect()
