+++ Projects/DonWorry/Sources/Home/Cells/TakeBillCardCollectionViewCell.swift
@@ -1,5 +1,5 @@
 //
-//  GivePaymentCardCollectionViewCell.swift
+//  TakeBillCardCollectionViewCell.swift
 //  DonWorry
 //
 //  Created by Woody on 2022/08/09.
@@ -8,49 +8,23 @@
 
 import UIKit
 import DesignSystem
-import DonWorryExtensions
-import Kingfisher
 
-struct GivePaymentCardCellViewModel: Equatable {
-    var takerID: Int
-    var imageURL: String
-    var nickName: String
+struct TakeBillCardCellViewModel: Equatable {
+    var giverID: Int
     var amount: String
     var isCompleted: Bool
 }
 
-final class GivePaymentCardCollectionViewCell: UICollectionViewCell {
+final class TakeBillCardCollectionViewCell: UICollectionViewCell {
 
-    lazy var wholeStackView: UIStackView = {
-        let v = UIStackView()
-        v.axis = .vertical
-        v.alignment = .center
-        v.distribution = .equalSpacing
-        return v
-    }()
-
-    lazy var profileStackView: UIStackView = {
-        let v = UIStackView()
-        v.axis = .vertical
-        v.alignment = .center
-        v.distribution = .fill
-        v.spacing = 10
-        return v
-    }()
-    lazy var profileImageView: UIImageView = {
-        let v = UIImageView()
-        v.layer.borderColor = UIColor.designSystem(.white)?.cgColor
-        v.layer.borderWidth = 1
-        v.contentMode = .scaleAspectFill
-        return v
-    }()
-    lazy var nickNameLabel: UILabel = {
+    lazy var titleLabel: UILabel = {
         let v = UILabel()
+        v.font = .designSystem(weight: .bold, size: ._13)
         v.textColor = .designSystem(.white)
-        v.font = .designSystem(weight: .heavy, size: ._15)
-        v.textAlignment = .center
+        v.text = "받을 돈"
         return v
     }()
+
     lazy var descriptionStackView: UIStackView = {
         let v = UIStackView()
         v.axis = .vertical
@@ -63,7 +37,7 @@ final class GivePaymentCardCollectionViewCell: UICollectionViewCell {
         let v = UILabel()
         v.textColor = .designSystem(.white)
         v.font = .designSystem(weight: .bold, size: ._13)
-        v.text = "줄돈"
+        v.text = "미정산"
         v.textAlignment = .center
         return v
     }()
@@ -80,14 +54,9 @@ final class GivePaymentCardCollectionViewCell: UICollectionViewCell {
         return v
     }()
 
-    var viewModel: GivePaymentCardCellViewModel? {
+    var viewModel: TakeBillCardCellViewModel? {
         didSet {
-            self.nickNameLabel.text = viewModel?.nickName
             self.amountLabel.text = viewModel?.amount
-            if let viewModel = viewModel {
-                let urlString = URL(string: viewModel.imageURL)
-                profileImageView.kf.setImage(with: urlString)
-            }
             if let viewModel = viewModel {
                 completeCoverView.isHidden = !viewModel.isCompleted
             }
@@ -106,29 +75,23 @@ final class GivePaymentCardCollectionViewCell: UICollectionViewCell {
     }
 
     private func setUI() {
-        self.contentView.backgroundColor = .designSystem(.brown)
-        self.contentView.addSubview(self.wholeStackView)
+        self.contentView.backgroundColor = .designSystem(.green)
         self.contentView.addSubview(self.completeCoverView)
-        self.wholeStackView.addArrangedSubview(self.profileStackView)
-        self.wholeStackView.addArrangedSubview(self.descriptionStackView)
-        self.profileStackView.addArrangedSubview(self.profileImageView)
-        self.profileStackView.addArrangedSubview(self.nickNameLabel)
+        self.contentView.addSubview(self.titleLabel)
+        self.contentView.addSubview(self.descriptionStackView)
         self.descriptionStackView.addArrangedSubview(self.descriptionLabel)
         self.descriptionStackView.addArrangedSubview(self.amountLabel)
 
-        self.wholeStackView.snp.makeConstraints { make in
-            make.top.equalToSuperview().offset(25)
-            make.bottom.equalToSuperview().inset(19)
-            make.leading.trailing.equalToSuperview().inset(25)
+        self.titleLabel.snp.makeConstraints { make in
+            make.top.equalToSuperview().offset(15)
+            make.leading.equalToSuperview().offset(14)
         }
-        self.profileImageView.snp.makeConstraints { make in
-            make.width.height.equalTo(60)
+        self.descriptionStackView.snp.makeConstraints { make in
+            make.center.equalToSuperview()
         }
         self.completeCoverView.snp.makeConstraints { make in
             make.edges.equalToSuperview()
         }
-
-        self.profileImageView.roundCorners(30)
         self.contentView.roundCorners(8)
         self.addShadow(shadowColor: UIColor.black.withAlphaComponent(0.4).cgColor)
         self.addCompleteCoverViewBlurEffect()
